fixed loading non challenge savegames and another crash
As every filename contained _ (due to the foldername of the saves
(.ffill_saves) the check didn't really work.

Also escape does not return the string entered but null - this is more
consistent with what the user expects i think.

also added a Console.ReadKey to the other checks so the user has time to
read the error message if the error occurs.